refactor: revert src/query/src/dummy_catalog.rs
