+++ src/catalog/src/lib.rs
@@ -27,7 +27,6 @@ use table::TableRef;
 
 use crate::error::Result;
 
-pub mod dummy_catalog;
 pub mod error;
 pub mod information_schema;
 pub mod kvbackend;
